feat: modularize React provider templates and add new tests
- Replaced `reactProviderComponentsTemplate` with new modular templates: `reactIntrigProviderTemplate`, `reactIntrigProviderStubTemplate`, and `reactStatusTrapTemplate`.
- Added detailed tests to validate the syntax and content of the new templates.
- Updated `react-binding.service.ts` and `main.template.ts` to include the new modularized templates.
- Improved maintainability and scalability of the provider structure with enhanced exports and TypeScript integration.
