+++ lib/react-binding/src/lib/templates/provider/intrig-provider.template.ts
@@ -1,21 +1,11 @@
-import {IntrigSourceConfig, typescript} from "common";
-import * as path from 'path'
+import { IntrigSourceConfig, typescript } from 'common';
+import * as path from 'path';
 
-export function reactProviderComponentsTemplate(_path: string, apisToSync: IntrigSourceConfig[]) {
-  const ts = typescript(path.resolve(_path, "src", "provider-components.tsx"))
-  return ts`import React, {
-  PropsWithChildren,
-  useCallback,
-  useContext,
-  useMemo,
-  useReducer,
-  useState,
-} from 'react';
+export function reactIntrigProviderTemplate(_path: string, _apisToSync: IntrigSourceConfig[]) {
+  const ts = typescript(path.resolve(_path, 'src', 'intrig-provider.tsx'));
+  return ts`import React, { useMemo, useReducer } from 'react';
 import {
   error,
-  isSuccess,
-  isError,
-  isPending,
   NetworkState,
   pending,
   success,
@@ -23,27 +13,14 @@ import {
   configError,
   httpError,
   networkError,
-  init,
 } from './network-state';
-import {
-  Axios,
-  AxiosProgressEvent,
-  AxiosResponse,
-  isAxiosError,
-} from 'axios';
+import { Axios, AxiosResponse, isAxiosError } from 'axios';
 import { ZodError, ZodSchema } from 'zod';
-import logger from './logger';
 import { flushSync } from 'react-dom';
 import { createParser } from 'eventsource-parser';
 
 import { Context, RequestType, GlobalState } from './intrig-context';
-import {
-  DefaultConfigs,
-  IntrigProviderProps,
-  IntrigProviderStubProps,
-  StatusTrapProps,
-  StubType,
-} from './interfaces';
+import { IntrigProviderProps } from './interfaces';
 import { createAxiosInstances } from './axios-config';
 import { requestReducer, inferNetworkReason } from './reducer';
 
@@ -51,10 +28,7 @@ import { requestReducer, inferNetworkReason } from './reducer';
  * IntrigProvider is a context provider component that sets up global state management
  * and provides Axios instances for API requests.
  */
-export function IntrigProvider({
-  children,
-  configs = {},
-}: IntrigProviderProps) {
+export function IntrigProvider({ children, configs = {} }: IntrigProviderProps) {
   const [state, dispatch] = useReducer(requestReducer, {} as GlobalState);
 
   const axiosInstances: Record<string, Axios> = useMemo(() => {
@@ -207,134 +181,5 @@ export function IntrigProvider({
 
   return <Context.Provider value={contextValue}>{children}</Context.Provider>;
 }
-
-export function IntrigProviderStub({
-  children,
-  configs = {},
-  stubs = () => {},
-}: IntrigProviderStubProps) {
-  const [state, dispatch] = useReducer(requestReducer, {} as GlobalState);
-
-  const collectedStubs = useMemo(() => {
-    let fns: Record<
-      string,
-      (
-        params: any,
-        body: any,
-        dispatch: (state: NetworkState<any>) => void,
-      ) => Promise<void>
-    > = {};
-    function stub<P, B, T>(
-      hook: any,
-      fn: (
-        params: P,
-        body: B,
-        dispatch: (state: NetworkState<T>) => void,
-      ) => Promise<void>,
-    ) {
-      fns[hook.key] = fn;
-    }
-    stubs(stub);
-    return fns;
-  }, [stubs]);
-
-  const contextValue = useMemo(() => {
-    async function execute<T>(
-      request: RequestType,
-      dispatch: (state: NetworkState<T>) => void,
-      schema: ZodSchema<T> | undefined,
-    ) {
-      let stub = collectedStubs[request.key];
-
-      if (!!stub) {
-        try {
-          await stub(request.params, request.data, dispatch);
-        } catch (e: any) {
-          dispatch(error(configError(e?.message ?? '')));
-        }
-      } else {
-        dispatch(init());
-      }
-    }
-
-    return {
-      state,
-      dispatch,
-      filteredState: state,
-      configs,
-      execute,
-    };
-  }, [state, dispatch, configs, collectedStubs]);
-
-  return <Context.Provider value={contextValue}>{children}</Context.Provider>;
-}
-
-/**
- * StatusTrap component is used to track and manage network request states.
- */
-export function StatusTrap({
-  children,
-  type,
-  propagate = true,
-}: PropsWithChildren<StatusTrapProps>) {
-  const ctx = useContext(Context);
-
-  const [requests, setRequests] = useState<string[]>([]);
-
-  const shouldHandleEvent = useCallback(
-    (state: NetworkState) => {
-      switch (type) {
-        case 'error':
-          return isError(state);
-        case 'pending':
-          return isPending(state);
-        case 'pending + error':
-          return isPending(state) || isError(state);
-        default:
-          return false;
-      }
-    },
-    [type],
-  );
-
-  const dispatch = useCallback(
-    (event: any) => {
-      if (!event.handled) {
-        if (shouldHandleEvent(event.state)) {
-          setRequests((prev) => [...prev, event.key]);
-          if (!propagate) {
-            ctx.dispatch({
-              ...event,
-              handled: true,
-            });
-            return;
-          }
-        } else {
-          setRequests((prev) => prev.filter((k) => k !== event.key));
-        }
-      }
-      ctx.dispatch(event);
-    },
-    [ctx, propagate, shouldHandleEvent],
-  );
-
-  const filteredState = useMemo(() => {
-    return Object.fromEntries(
-      Object.entries(ctx.state).filter(([key]) => requests.includes(key)),
-    );
-  }, [ctx.state, requests]);
-
-  return (
-    <Context.Provider
-      value={{
-        ...ctx,
-        dispatch,
-        filteredState,
-      }}
-    >
-      {children}
-    </Context.Provider>
-  );
-}
-  `
+`;
 }
